app/requirements.txt
# Core - latest versions
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=21.2.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.17
httpx>=0.28.0

# Database
sqlalchemy>=2.0.36
alembic>=1.12.0
psycopg2-binary>=2.9.10

# Auth (JWT, password hashing)
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0,<5.0.0

# Validation (EmailStr in Pydantic)
email-validator>=2.2.0

# AI - google-genai (unified SDK, replaces deprecated google-generativeai)
google-genai>=1.62.0
numpy>=1.26.0

# Graph
neo4j>=5.25.0

# Cache
redis>=5.2.0

# MCP - Model Context Protocol for connecting to Google MCP servers
mcp>=1.12.0

# STIX 2.1 - OASIS cyber threat intelligence standard for government SIEM interoperability
stix2>=3.0.0

# Optional heavy ML deps (install if you want local inference)
torch>=2.2.0
transformers>=4.40.0
sentencepiece>=0.2.0
onnxruntime>=1.18.0
onnx>=1.16.0
pgvector==0.2.4
sentence-transformers==2.2.2

# Image processing
pillow>=10.0.0
